refactoring(Area) : deplace le code specific dans une classe AreaSystem
