+++ Systems/DialogSystem/DialogSystem.cs
@@ -177,8 +177,8 @@ namespace NWN.Systems
       if (Players.TryGetValue(NWScript.GetLastSpeaker(), out player))
       {
         Area area; 
-        if (Module.areaDictionnary.TryGetValue(NWScript.GetObjectUUID(NWScript.GetArea(player.oid)), out area))
-          area.StartEntryScene(player);
+        if (AreaSystem.areaDictionnary.TryGetValue(NWScript.GetObjectUUID(NWScript.GetArea(player.oid)), out area))
+          AreaSystem.StartEntryScene(area, player);
       }
 
       return 0;
